fixed android build

+++ 3rdparty/untz/include/UntzSound.h
@@ -13,7 +13,7 @@
 
 namespace UNTZ
 {
-	typedef Int64 (StreamCallback)(float* buffers, UInt32 numChannels, UInt32 length, void* userdata);
+	typedef UInt32 (StreamCallback)(float* buffers, UInt32 numChannels, UInt32 length, void* userdata); 
 
 	typedef struct {
 		UInt32 mBitsPerSample;
@@ -58,7 +58,6 @@ namespace UNTZ
 		bool isPaused();
 
 		SoundData* getData() const { return mpData; }
-		
 	private:
 		Sound();
 		SoundData* mpData;
